further refactoring of tests
