+++ cypress/e2e/pages/viewApplication.cy.ts
@@ -7,6 +7,26 @@ import { StatusCodes } from 'http-status-codes';
 
 const applicationId = faker.string.uuid();
 const personId = faker.string.uuid();
+const response = {
+  id: 'b0058e24-ef9c-4672-a830-dc5c1a329dce',
+  targetId: applicationId,
+  type: 'update',
+  targetType: 'housingApplication',
+  createdAt: '2024-09-04T13:13:59.7981926Z',
+  timetoLiveForRecord: 0,
+  oldData: {
+    'assessment.biddingNumber': null,
+  },
+  newData: {
+    _activityType: 'biddingNumberChangedByUser',
+    'assessment.biddingNumber': '1000001',
+  },
+  authorDetails: {
+    fullName: 'John Jones',
+    email: 'test@hackney.gov.uk',
+  },
+  sourceDomain: 'HousingRegister',
+};
 
 describe('View a resident application', () => {
   beforeEach(() => {
@@ -20,7 +40,7 @@ describe('View a resident application', () => {
     application.status = ApplicationStatus.SUBMITTED;
 
     cy.loginAsUser('readOnly');
-    cy.mockActivityHistoryApiEmptyResponse(applicationId);
+    cy.mockActivityHistoryApiEmptyResponse(applicationId, response);
     cy.mockHousingRegisterApiGetApplications(applicationId, application);
 
     ViewApplicationPage.visit(applicationId);
@@ -80,7 +100,7 @@ describe('View a resident application', () => {
     cy.mockHousingRegisterApiGetApplications(applicationId, application);
 
     ViewApplicationPage.visit(applicationId);
-    ViewApplicationPage.getNotesAndHistorySection().should('exist');
+    ViewApplicationPage.getNavItem('history').should('exist');
   });
 
   it('allows readonly users to see note section', () => {
@@ -89,10 +109,10 @@ describe('View a resident application', () => {
     application.status = ApplicationStatus.SUBMITTED;
 
     cy.loginAsUser('readOnly');
-    cy.mockActivityHistoryApiEmptyResponse(applicationId, true);
+    cy.mockActivityHistoryApiEmptyResponse(applicationId, response, true);
     cy.mockHousingRegisterApiGetApplications(applicationId, application, true);
     ViewApplicationPage.visit(applicationId);
-    ViewApplicationPage.getNotesAndHistorySection()
+    ViewApplicationPage.getNavItem('history')
       .click()
       .get('[data-testid="test-add-text-input"]')
       .should('exist');
@@ -104,15 +124,15 @@ describe('View a resident application', () => {
     application.status = ApplicationStatus.SUBMITTED;
 
     cy.loginAsUser('readOnly');
-    cy.mockActivityHistoryApiEmptyResponse(applicationId, true);
+    cy.mockActivityHistoryApiEmptyResponse(applicationId, response, true);
     cy.mockHousingRegisterApiGetApplications(applicationId, application, true);
 
     ViewApplicationPage.visit(applicationId);
-    ViewApplicationPage.getNotesAndHistorySection()
+    ViewApplicationPage.getNavItem('history')
       .click()
       .get('[data-testid="test-activity-history"]')
       .should('exist');
-    ViewApplicationPage.getNotesAndHistorySection()
+    ViewApplicationPage.getNavItem('history')
       .click()
       .get('[data-testid="test-detailed-history-items"]')
       .should('not.exist');
